add properties

+++ test-clients/src/main/java/com/hedera/services/bdd/spec/transactions/schedule/HapiScheduleCreate.java
@@ -59,8 +59,8 @@ public class HapiScheduleCreate<T extends HapiTxnOp<T>> extends HapiTxnOp<HapiSc
 
     private static final int defaultScheduleTxnExpiry =
             HapiSpecSetup.getDefaultNodeProps().getInteger("ledger.schedule.txExpiryTimeSecs");
-    private static final int defaultScheduleTxCostIncrementUSD =
-            HapiSpecSetup.getDefaultNodeProps().getInteger("scheduling.costIncrementUSD");
+    private static final double defaultScheduleTxCostIncrementUSD =
+            HapiSpecSetup.getDefaultNodeProps().getDouble("scheduling.costIncrementUSD");
     private static final int defaultScheduleTxCostIncrementBytesPerMonth =
             HapiSpecSetup.getDefaultNodeProps().getInteger("scheduling.costIncrementBytesPerMonth");
